Creting UIColor BeerColor extension, and add americanWheat and americanPaleAle colors
